requirements.txt
Flask
pandas
plotly
kaleido
python-docx
reportlab
gunicorn
